seperated the linux and windows process scanning into different files
